Fix precision bug caused by move to microseconds for timewarp calc.

+++ src/main/java/org/eclipselabs/garbagecat/util/jdk/JdkUtil.java
@@ -846,7 +846,7 @@ public final class JdkUtil {
             durationThreshold = durationThreshold.movePointLeft(2);
             durationThreshold = durationThreshold.multiply(new BigDecimal(interval));
             durationThreshold.setScale(0, RoundingMode.DOWN);
-            return gcEvent.getDuration() > durationThreshold.intValue();
+            return gcEvent.getDuration() > durationThreshold.longValue();
         }
     }
